+++ eta-meeting-organizer-backend/src/main/java/hu/flowacademy/meetingorganizer/persistance/model/User.java
@@ -1,4 +1,4 @@
-package hu.flowacademy.meetingorganizer.persistence.model;
+package hu.flowacademy.meetingorganizer.persistance.model;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
