+++ eta-meeting-organizer-backend/src/main/java/hu/flowacademy/meetingorganizer/service/UserService.java
@@ -1,6 +1,6 @@
 package hu.flowacademy.meetingorganizer.service;
-import hu.flowacademy.meetingorganizer.persistence.model.User;
-import hu.flowacademy.meetingorganizer.persistence.repository.UserRepository;
+import hu.flowacademy.meetingorganizer.persistance.model.User;
+import hu.flowacademy.meetingorganizer.persistance.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -9,7 +9,6 @@ import org.springframework.web.server.ResponseStatusException;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Transactional
